j'ai pas encore finis la page mais j'ai presque finis de rendre cette partie responsive pour les peties ecrants

+++ Joss/convertion_vw et vh.py
@@ -1,7 +1,7 @@
-px =int(input("entre la dimention a convertir" ))
+px =float(input("entre la dimention a convertir" ))
 unite= input("entrer l'unite vers la quel la dimention doit etre convert")
 
-def convertir(px:int,unite:chr):
+def convertir(px:float,unite:chr):
 
     if unite == "vw":
